+++ src/cast_core/include/cast_core/actors/RunCommand.hpp
@@ -19,8 +19,6 @@
 #include <gennylib/PhaseLoop.hpp>
 #include <gennylib/context.hpp>
 
-#include <cast_core/config/RunCommandConfig.hpp>
-
 namespace genny::actor {
 
 /**
